# life/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Life Pratics project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
        "/especialidades": {
            "get": {
                "tags": [
                    "Especialidades",
                    "Admin"
                ],
                "summary": "Exibe uma lista de especialidades",
                "operationId": "1ab85db17a14c74ae0522af66aab8704",
                "responses": {
                    "200": {
                        "description": "Lista de especialidades"
                    }
                }
            },
            "post": {
                "tags": [
                    "Especialidades",
                    "Admin"
                ],
                "summary": "Armazena uma nova especialidade",
                "operationId": "475d523f2544e35785500f3e694f08bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Cardiologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Especialidade criada com sucesso"
                    }
                }
            }
        },
        "/especialidades/{id}": {
            "get": {
                "tags": [
                    "Especialidades",
                    "Admin"
                ],
                "summary": "Exibe uma especialidade específica",
                "operationId": "d33d924147f7344e8a0bd41fdd856dfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da especialidade"
                    }
                }
            },
            "put": {
                "tags": [
                    "Especialidades",
                    "Admin"
                ],
                "summary": "Atualiza uma especialidade específica",
                "operationId": "de0e4e120e10685d5d0bd82258f805c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Neurologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Especialidade atualizada com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Especialidades",
                    "Admin"
                ],
                "summary": "Remove uma especialidade específica",
                "operationId": "fa845b8904abf331a82366579bc66f8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialidade removida com sucesso"
                    }
                }
            }
        },
        "/medico": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lista todos os médicos",
                "operationId": "9c090414fba5e409412f3f4f97768b68",
                "responses": {
                    "200": {
                        "description": "Lista de médicos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Cria um novo médico",
                "operationId": "f04e3c7cab5f3aea85a714159f994ce0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "med_crm": {
                                        "type": "string"
                                    },
                                    "med_nome": {
                                        "type": "string"
                                    },
                                    "med_email": {
                                        "type": "string"
                                    },
                                    "med_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Médico criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/medico/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Exibe um médico específico",
                "operationId": "45ba76bb99e282edec1fd073241c417c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do médico"
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Atualiza um médico existente",
                "operationId": "af611b4c8109326db2e233b25bb27d2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "med_crm": {
                                        "type": "string"
                                    },
                                    "med_nome": {
                                        "type": "string"
                                    },
                                    "med_email": {
                                        "type": "string"
                                    },
                                    "med_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Médico atualizado com sucesso"
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remove um médico",
                "operationId": "d57752c7065c4d05b245f7f141124aca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Médico removido com sucesso"
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                }
            }
        },
...
        "/medico/consultas/{consulta_id}": {
            "get": {
                "tags": [
                    "Medico"
                ],
                "summary": "Exibe uma consulta específica do médico autenticado",
                "operationId": "fcdc5a15b7adfe3939eaa38285b35631",
                "parameters": [
                    {
                        "name": "consulta_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da consulta"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                }
            }
        },
...
        "/medico/pacientes/{paciente_id}": {
            "get": {
                "tags": [
                    "Medico"
                ],
                "summary": "Exibe um paciente específico do médico autenticado",
                "operationId": "3b9ff9fd5a1e0c9e5d2bea695d3c353d",
                "parameters": [
                    {
                        "name": "paciente_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do paciente"
                    },
                    "404": {
                        "description": "Paciente não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}